Reasoner related functions are moved to static funcs
